backend/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
boto3==1.34.69
botocore==1.34.69
aiofiles==23.2.1
sse-starlette==2.0.0
faiss-cpu==1.7.4
chromadb==0.5.0
numpy==1.26.4
scikit-learn==1.4.2